Refactored the login view to display on the StartModal
